Generador de imagenes/requirements.txt
accelerate==1.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
diffusers==0.34.0
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
psutil==7.0.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
urllib3==2.5.0
zipp==3.23.0
